adapt journal completed job to handle wiki content
